Move repair functions to separate module
